Update fingerprinted CSS path
The CSS file has a hash/fingerprint in it that it dynamically generated
when the website is updated. Since this playground is effectively a
modified clone of the generated site, it's gotten out of date with the
fingerprint, and the styling isn't being applied properly. This brings
it up to date

https://github.com/dxw/website/blob/c7e67e3583c82a069ec0059107ff468d2bf43921/wp-content/themes/dxw-child-theme/app/Theme/Scripts.php#L40